add integrandfunction to exponent

+++ SSB/src/ss/week4/math/Exponent.java
@@ -34,7 +34,7 @@ public class Exponent implements Function, Integrandable {
 	
 	@Override
 	public Function integrand() {
-		return new LinearProduct(new Constant(1 / (n + 1)), new Exponent(n + 1));
+		return new LinearProduct(new Constant(1 / (double) (n + 1)), new Exponent(n + 1));
 	}
 
 }
